requirements.txt
fastapi
uvicorn
httpx
python-dotenv
textblob
google-generativeai
